+++ .eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = {
+  env: {
+    commonjs: true,
+    es2021: true,
+    node: true
+  },
+  extends: 'standard-with-typescript',
+  overrides: [
+  ],
+  parserOptions: {
+    ecmaVersion: 'latest'
+  },
+  rules: {
+  }
+}
