--- .eslintrc.json
@@ -1,13 +1,0 @@
-{
-  "env": {
-    "es2021": true,
-    "node": true
-  },
-  "extends": [
-    "standard"
-  ],
-  "parserOptions": {
-    "ecmaVersion": "latest"
-  },
-  "rules": {}
-}
