+++ package.json
@@ -2,7 +2,7 @@
   "name": "alpaca.js",
   "version": "0.2.0",
   "description": "Logging for Node.js ✍️  with a opinionated theme 🎨  and configurable API  🍭",
-  "main": "index.js",
+  "main": "dist/index.js",
   "type": "commonjs",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -30,12 +30,19 @@
     "winston": "^3.7.2"
   },
   "devDependencies": {
+    "@types/node": "^18.8.3",
+    "@types/node-emoji": "^1.8.2",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
     "eslint": "^8.0.1",
     "eslint-config-standard": "^17.0.0",
+    "eslint-config-standard-with-typescript": "^23.0.0",
     "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-n": "^15.0.0",
     "eslint-plugin-promise": "^6.0.0",
     "prettier": "^2.6.2",
-    "prettier-config-standard": "^5.0.0"
-  }
+    "prettier-config-standard": "^5.0.0",
+    "tsc": "^2.0.4",
+    "typescript": "*"
+  },
+  "types": "./dist/index.d.ts"
 }
